requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
distro==1.9.0
Django==5.0.3
dotenv==0.19.0
exceptiongroup==1.2.0
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
jsonpath-python==1.0.6
marshmallow==3.21.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
retell==0.5.1
retell-sdk==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
djangorestframework==3.15.1